refactor(katana-primitives): move class related types to a separate module (#1560)
