+++ crates/katana/primitives/src/genesis/mod.rs
@@ -24,10 +24,8 @@ use self::constant::{
     OZ_ACCOUNT_CONTRACT_PUBKEY_STORAGE_SLOT,
 };
 use crate::block::{Block, BlockHash, BlockNumber, GasPrices, Header};
-use crate::contract::{
-    ClassHash, CompiledClass, CompiledClassHash, ContractAddress, FlattenedSierraClass, StorageKey,
-    StorageValue,
-};
+use crate::class::{ClassHash, CompiledClass, CompiledClassHash, FlattenedSierraClass};
+use crate::contract::{ContractAddress, StorageKey, StorageValue};
 use crate::state::StateUpdatesWithDeclaredClasses;
 use crate::utils::split_u256;
 use crate::version::CURRENT_STARKNET_VERSION;
