+++ crates/katana/primitives/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod block;
 pub mod chain;
+pub mod class;
 pub mod contract;
 pub mod env;
 pub mod event;
