+++ crates/katana/storage/db/src/models/contract.rs
@@ -1,5 +1,6 @@
 use katana_primitives::block::BlockNumber;
-use katana_primitives::contract::{ClassHash, ContractAddress, Nonce};
+use katana_primitives::class::ClassHash;
+use katana_primitives::contract::{ContractAddress, Nonce};
 use serde::{Deserialize, Serialize};
 
 use crate::codecs::{Compress, Decode, Decompress, Encode};
